% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev_function}
\alias{rev_function}
\title{Review Function
This function allows you to find out about reviewer 2's thoughts about your paper.}
\usage{
rev_function(review = TRUE)
}
\arguments{
\item{review}{Have you completed your paper submission? Defaults to TRUE.}
}
\description{
Review Function
This function allows you to find out about reviewer 2's thoughts about your paper.
}
\examples{
rev_function()
}
\keyword{paper}
